feat: seed 400 guests and ~750 proposals in dev via a large profile
Dev pages now show realistic volumes so slow list rendering surfaces
before a real event hits it. The bulk layer is generated from word
lists + a seeded RNG (no fixture dump to maintain, dates stay relative
to today) on top of the unchanged curated fixtures; E2E always seeds
the curated-only small profile, so tests stay fast and cannot come to
depend on bulk data. Seed script split into scripts/seed/ modules:
curated data under data/, generator in bulk.ts.
