+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The schedule, the session and proposal forms and the admin locations page now name colours by role instead of by palette shade. A location's colour is derived from its hue with `color-mix()` against the current surface and foreground, so a single `loc-<name>` class works in both themes and the `@source inline(...)` safelist is gone; `tests/unit/location-colors.test.ts` recomputes the contrast of all 22 hues in both themes. The rest of the app still uses palette shades
 - Dependabot now keeps the mailpit image up to date. Its pin was repeated in `docker-compose.dev.yml` and in both CI workflows, and Dependabot cannot see a workflow's `services:` image ([dependabot-core#5819](https://github.com/dependabot/dependabot-core/issues/5819)), so CI now starts mailpit from the compose file and the pin has a single home. The image carries a tag as well as a digest, since a digest alone gives Dependabot no version to compare
 - Seeded guests with a profile now carry a last-edited date, spread over the past month, instead of all being undated — the seed writes guest rows directly, so nothing stamped them. "Recently updated" had nothing to sort in dev and demos
+- The dev seed (`make dev-db-seed`) now defaults to a `large` profile: 400 guests, ~750 proposals and a mostly filled Gamma schedule, generated deterministically from word lists on top of the unchanged hand-curated fixtures — so dev pages show realistic volumes and slow list rendering surfaces before a real event hits it. The E2E suite always seeds the curated-only `small` profile (`SEED_PROFILE` selects one explicitly), so tests stay fast and can never come to depend on bulk data. The seed script is split into `scripts/seed/` modules: curated data under `data/`, the bulk generator in `bulk.ts`
 - The seed data gives Conference Gamma four more proposals: two hosted by Hana Kobayashi, whose profile is the most complete of the seeded guests, and two with no host yet. Both cases were previously only reachable through the random host assignment, which made them awkward to point a screenshot at
 
 ## [3.3.1] - 2026-08-08
