+++ tests/e2e/init.ts
@@ -1,10 +1,11 @@
-import { resetDatabase } from "@/scripts/seed-database";
+import { resetDatabase } from "@/scripts/seed/seed-database";
 
 async function globalSetup() {
   console.log("🚀 Setting up test environment...");
 
-  // Reset and seed the database with test data
-  await resetDatabase();
+  // Reset and seed the database with test data. Always the small profile:
+  // the suite pins on the curated fixtures and must never see bulk data.
+  await resetDatabase("small");
 
   console.log("✅ Test environment ready!");
 }
